// src/enums/WebRoutes.ts
export enum WebRoutes {
  HOME = "/",
  LOGIN = "/auth/login",
  SIGNUP = "/auth/signup",
  FORGOT_PW = "/auth/forgot",
  RESET_PW = "/auth/reset",
  SETTINGS = "/settings",
  SINGLE_PRODUCT = "/products/:pid",
  MY_PROFILE = "/settings/myprofile",
  MY_TRANSACTIONS = "/settings/transactions",
  MY_Address = "/settings/myaddress",
  CHANGE_PW = "/settings/changepassword",
  HELP_SUPPORT = "/settings/help",
  DASHBOARD = "/dashboard",
  DASH_SUMMARY = "/dashboard/summary",
  DASH_PRODUCTS = "/dashboard/products",
  DASH_ADD_PRODUCT = "/dashboard/products/add",
  DASH_EDIT_PRODUCT = "/dashboard/products/edit/:pid",
  DASH_ORDERS = "/dashboard/orders",
}